requirements.txt
# Core dependencies
PyGithub>=1.55
requests>=2.28.0

# AI providers - uncomment what you need
# Uncomment the providers you want to use:
google-generativeai>=0.3.0
# openai>=0.27.0
# anthropic>=0.18.0